requirements.txt
# Core AgentScope dependencies
agentscope>=1.0.0

# FastAPI and web framework
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.5.0,<3.0.0
gradio==4.44.0

# Database and caching
sqlalchemy[asyncio]>=2.0.0
alembic>=1.10.0
redis>=4.5.0
asyncpg>=0.29.0  # PostgreSQL async driver
psycopg2-binary>=2.9.0  # PostgreSQL driver for SQLAlchemy

# Async task processing
celery>=5.3.0
kombu>=5.3.0

# Additional utilities
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
shortuuid>=1.0.0
aiofiles>=23.0.0
python-dotenv>=1.0.0

# Monitoring and logging
# structlog>=23.0.0
prometheus-client>=0.16.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
httpx>=0.24.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Optional model providers (will be configured based on user needs)
openai>=1.0.0
anthropic>=0.3.0
dashscope>=1.0.0
ollama>=0.1.7